+++ code/go/0chain.net/smartcontract/storagesc/sc.go
@@ -228,7 +228,7 @@ func (sc *StorageSmartContract) Execute(t *transaction.Transaction,
 	case "add_validator":
 		resp, err = sc.addValidator(t, input, balances)
 	case "blobber_health_check":
-		resp, err = sc.blobberHealthCheck(t, input, balances)
+		resp, err = sc.blobberHeaelthCheck(t, input, balances)
 	case "validator_health_check":
 		resp, err = sc.validatorHealthCheck(t, input, balances)
 	case "update_blobber_settings":
